pyproject.toml
[project]
name = "lightrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.0.1",
    "aioboto3>=13.2.0",
    "aiohttp>=3.10.10",
    "graspologic>=3.4.1",
    "hnswlib>=0.8.0",
    "jsonlines>=4.0.0",
    "nano-vectordb>=0.0.4",
    "neo4j>=5.26.0",
    "nest-asyncio>=1.6.0",
    "networkx>=3.4.1",
    "ollama>=0.3.3",
    "openai>=1.52.0",
    "python-dotenv>=1.0.1",
    "pyvis>=0.3.2",
    "tenacity>=9.0.0",
    "tiktoken>=0.8.0",
    "torch>=2.5.0",
    "transformers>=4.45.2",
    "xxhash>=3.5.0",
]
